% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getAtomicWeight}
\alias{getAtomicWeight}
\title{getAtomicWeight}
\usage{
getAtomicWeight(element)
}
\arguments{
\item{element}{A chemical symbol for the element}
}
\value{
A data.frame with the standard atomic weight (AW) and its standard uncertainty (u_AW)
}
\description{
This function return the standard atomic weight and its uncertainty for a given element.
The standard atomic weights are calculated using the reported interval values for  normal and naturally  occurring materials
according to the 2019 Table of Standard Atomic Weights published at \url{https://ciaaw.org/atomic-weights.htm} (1), based on
IUPAC Technical Report: Atomic weights of the elements 2013(2). The Commission on Isotopic Abundances and Atomic Weights (CIAAW)
defines a normal material as any terrestrial material
that "is a reasonably possible source for this element or its compounds in commerce, for industry or science;
the material is not itself studied for some extraordinary anomaly and its isotopic composition has not been
modified significantly in a geologically brief period"(3).
}
\details{
This function takes the more simple aproximation for standard uncertaninty calculation, based on the
reported interval with the asumption of a rectangular distribution. Decisional uncertainty is returned when is reported(3).
More detailed uncertainty estimations considering isotopic abundances for an specific application requires more complex models such as the reported in (2-4)
}
\examples{
getAtomicWeight("H")
}
\references{
(1) CIAAW, Standard atomic weights \link{online}, \url{https://ciaaw.org/atomic-weights.htm}. Retrieved: 2021-05-05

(2) Juris Meija, et. al., Atomic weights of the elements 2013 (IUPAC Technical Report), Pure Appl. Chem. 2016; 88(3): 265–291. \url{https://doi.org/10.1515/pac-2015-0305}

(3) Possolo, A.,  van der Veen, A.,   Meija, J.,  Hibbert, D.B., Interpreting and propagating the uncertainty of the standard atomic weights,  Pure Appl. Chem. 2018; 90(2): 395–424.\url{https://doi.org/10.1515/pac-2016-0402}

(4) Juris Meija and Zoltan Mester. Atomic weight uncertainty calculation from isotopic composition of the elements, Metrologia 45 (2008) 459–463.\url{https://doi.org/10.1088/0026-1394/45/4/012}
}
\seealso{
\code{\link[=getMolarMass]{getMolarMass()}}
}
