% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getMolarMass}
\alias{getMolarMass}
\title{getMolarMass}
\usage{
getMolarMass(atomsType, atomsNumber = 1)
}
\arguments{
\item{atomsType}{A string vector with chemical symbols for the elements in the molecular formula}

\item{atomsNumber}{A int vector with the number of atoms for each element in the molecular formula (by default =1)}
}
\value{
A data.frame with the molar mass (MM) and its standard uncertainty (u_MM)
}
\description{
This function return the standard  molar mass and its standard uncertainty for a given molecular formula.
The standard atomic weights used are obtained from \code{\link[=getAtomicWeight]{getAtomicWeight()}}
}
\details{
The function uses the GUM aproach based on the molecular formula(1).
}
\examples{
getMolecularMass(c("C","H"),c(1,4))
}
\references{
(1) Possolo, A.,  van der Veen, A.,   Meija, J.,  Hibbert, D.B., Interpreting and propagating the uncertainty of the standard atomic weights,  Pure Appl. Chem. 2018; 90(2): 395–424.\url{https://doi.org/10.1515/pac-2016-0402}
}
\seealso{
\code{\link[=getAtomicWeights]{getAtomicWeights()}}
}
